Migrate tiledb_fragment_info_t struct definition into CAPI folder.
